+++ forestplot/plot.py
@@ -5,34 +5,34 @@ import numpy as np
 np.seterr(all="ignore")
 import matplotlib.pyplot as plt
 from matplotlib.pyplot import Axes
-from pyforestplot.arg_validators import check_data
-from pyforestplot.text_utils import form_est_ci
-from pyforestplot.text_utils import normalize_varlabels
-from pyforestplot.text_utils import indent_nongroupvar
-from pyforestplot.text_utils import format_varlabels
-from pyforestplot.text_utils import star_pval
-from pyforestplot.text_utils import prep_annote
-from pyforestplot.text_utils import prep_rightannnote
-from pyforestplot.text_utils import make_tableheaders
-from pyforestplot.dataframe_utils import insert_groups
-from pyforestplot.dataframe_utils import sort_groups
-from pyforestplot.dataframe_utils import sort_data
-from pyforestplot.dataframe_utils import reverse_dataframe
-from pyforestplot.graph_utils import right_flush_yticklabels
-from pyforestplot.graph_utils import draw_pval_right
-from pyforestplot.graph_utils import draw_ylabel1
-from pyforestplot.graph_utils import format_xlabel
-from pyforestplot.graph_utils import format_tableheader
-from pyforestplot.graph_utils import draw_yticklabel2
-from pyforestplot.graph_utils import remove_ticks
-from pyforestplot.graph_utils import format_grouplabels
-from pyforestplot.graph_utils import draw_ci
-from pyforestplot.graph_utils import draw_est_markers
-from pyforestplot.graph_utils import draw_ref_xline
-from pyforestplot.graph_utils import despineplot
-from pyforestplot.graph_utils import format_xticks
-from pyforestplot.graph_utils import draw_alt_row_colors
-from pyforestplot.graph_utils import draw_tablelines
+from forestplot.arg_validators import check_data
+from forestplot.text_utils import form_est_ci
+from forestplot.text_utils import normalize_varlabels
+from forestplot.text_utils import indent_nongroupvar
+from forestplot.text_utils import format_varlabels
+from forestplot.text_utils import star_pval
+from forestplot.text_utils import prep_annote
+from forestplot.text_utils import prep_rightannnote
+from forestplot.text_utils import make_tableheaders
+from forestplot.dataframe_utils import insert_groups
+from forestplot.dataframe_utils import sort_groups
+from forestplot.dataframe_utils import sort_data
+from forestplot.dataframe_utils import reverse_dataframe
+from forestplot.graph_utils import right_flush_yticklabels
+from forestplot.graph_utils import draw_pval_right
+from forestplot.graph_utils import draw_ylabel1
+from forestplot.graph_utils import format_xlabel
+from forestplot.graph_utils import format_tableheader
+from forestplot.graph_utils import draw_yticklabel2
+from forestplot.graph_utils import remove_ticks
+from forestplot.graph_utils import format_grouplabels
+from forestplot.graph_utils import draw_ci
+from forestplot.graph_utils import draw_est_markers
+from forestplot.graph_utils import draw_ref_xline
+from forestplot.graph_utils import despineplot
+from forestplot.graph_utils import format_xticks
+from forestplot.graph_utils import draw_alt_row_colors
+from forestplot.graph_utils import draw_tablelines
 from matplotlib import rcParams
 
 rcParams["font.monospace"] = [
