--- pyforestplot/__init__.py
@@ -1,2 +1,0 @@
-from pyforestplot.plot import forestplot
-from pyforestplot.dataframe_utils import load_data
